+++ service/gitlab/gitlab_mr_discussion.go
@@ -104,6 +104,11 @@ func (g *MergeRequestDiscussionCommenter) postCommentsForEach(ctx context.Contex
 		loc := c.Result.Diagnostic.GetLocation()
 		lnum := int(loc.GetRange().GetStart().GetLine())
 		body := commentutil.MarkdownComment(c)
+
+		if suggestion := commentutil.MarkdownSuggestion(c); suggestion != "" {
+			body = body + suggestion
+		}
+
 		if !c.Result.InDiffFile || lnum == 0 || postedcs.IsPosted(c, lnum, body) {
 			continue
 		}